refactor - split up tests
