+++ expo/tests/gbe/reports/test_staff_area.py
@@ -1,0 +1,61 @@
+from django.core.urlresolvers import reverse
+from django.test import TestCase, Client
+from tests.functions.gbe_functions import (
+    grant_privilege,
+    login_as,
+)
+from tests.factories.gbe_factories import (
+    ProfileFactory,
+    ShowFactory,
+)
+from gbe.models import Conference
+from tests.contexts import VolunteerContext
+
+
+class TestReviewStaffArea(TestCase):
+    def setUp(self):
+        self.client = Client()
+        self.profile = ProfileFactory()
+
+    def test_staff_area_path(self):
+        '''staff_area view should load
+        '''
+        show = ShowFactory()
+        context = VolunteerContext(event=show)
+        grant_privilege(self.profile, 'Act Reviewers')
+        login_as(self.profile, self)
+        response = self.client.get(
+            reverse('staff_area',
+                    urlconf="gbe.reporting.urls",
+                    args=[show.eventitem_id]))
+        self.assertEqual(response.status_code, 200)
+
+    def test_staff_area_with_inactive(self):
+        '''staff_area view should load
+        '''
+        show = ShowFactory()
+        inactive = ProfileFactory(
+            display_name="DON'T SEE THIS",
+            user_object__is_active=False
+        )
+        context = VolunteerContext(event=show, profile=inactive)
+        grant_privilege(self.profile, 'Act Reviewers')
+        login_as(self.profile, self)
+        response = self.client.get(
+            reverse('staff_area',
+                    urlconf="gbe.reporting.urls",
+                    args=[show.eventitem_id]))
+        self.assertEqual(response.status_code, 200)
+        self.assertTrue(
+            '<tr style="color:red;">' in response.content)
+
+    def test_staff_area_path_fail(self):
+        '''staff_area view should fail for non-authenticated users
+        '''
+        show = ShowFactory()
+        login_as(self.profile, self)
+        response = self.client.get(
+            reverse('staff_area',
+                    urlconf="gbe.reporting.urls",
+                    args=[show.eventitem_id]))
+        self.assertEqual(response.status_code, 403)
